CollegeManagementSystem/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
PyJWT==2.8.0